services/api/requirements.txt
# Lesson Plan Generator API

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
firebase-admin>=6.4.0
openai>=1.10.0
httpx>=0.26.0
